+++ src/media/midi/mod.rs
@@ -1,0 +1,25 @@
+// devela::media::midi
+//
+//! Midi functionality.
+//
+// safety
+// #![cfg_attr(feature = "safe_midi", forbid(unsafe_code))]
+
+mod error;
+
+crate::items! { // structural access: _mods, _all, _always
+    #[allow(unused)]
+    pub use _mods::*;
+    #[allow(unused)] #[doc(hidden)] #[doc(no_inline)]
+    pub use _always::*;
+
+    mod _mods {
+        pub use super::error::*;
+    }
+    pub(super) mod _all {
+        #[doc(inline)]
+        pub use super::_mods::*;
+    }
+    pub(super) mod _always { #![allow(unused)]
+    }
+}